+++ draft-packages/button/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.1.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@6.0.7...@kaizen/draft-button@6.1.0) (2022-05-20)
+
+
+### Features
+
+* add compatibility for react-18 ([#2731](https://github.com/cultureamp/kaizen-design-system/issues/2731)) ([0051f4c](https://github.com/cultureamp/kaizen-design-system/commit/0051f4cee82895acc2c2f44fc7bf8063857de57e))
+
+
+
+
+
 ## [6.0.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-button@6.0.6...@kaizen/draft-button@6.0.7) (2022-05-19)
 
 **Note:** Version bump only for package @kaizen/draft-button