+++ packages/a11y/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.29...@kaizen/a11y@1.4.0) (2022-05-20)
+
+
+### Features
+
+* add compatibility for react-18 ([#2731](https://github.com/cultureamp/kaizen-design-system/issues/2731)) ([0051f4c](https://github.com/cultureamp/kaizen-design-system/commit/0051f4cee82895acc2c2f44fc7bf8063857de57e))
+
+
+
+
+
 ## [1.3.29](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/a11y@1.3.28...@kaizen/a11y@1.3.29) (2022-05-19)
 
 **Note:** Version bump only for package @kaizen/a11y